refactor: prefer use of `getCustomConfig` to access custom config, also move to `server/services/Config`
