+++ api/server/services/Config/getCustomConfig.js
@@ -1,6 +1,6 @@
 const { CacheKeys } = require('librechat-data-provider');
-const loadCustomConfig = require('~/server/services/Config/loadCustomConfig');
-const getLogStores = require('./getLogStores');
+const loadCustomConfig = require('./loadCustomConfig');
+const getLogStores = require('~/cache/getLogStores');
 
 /**
  * Retrieves the configuration object
