+++ api/server/services/Config/loadConfigEndpoints.js
@@ -1,18 +1,12 @@
-const { CacheKeys, EModelEndpoint } = require('librechat-data-provider');
+const { EModelEndpoint } = require('librechat-data-provider');
 const { isUserProvided, extractEnvVariable } = require('~/server/utils');
-const loadCustomConfig = require('./loadCustomConfig');
-const { getLogStores } = require('~/cache');
+const getCustomConfig = require('./getCustomConfig');
 
 /**
  * Load config endpoints from the cached configuration object
  * @function loadConfigEndpoints */
 async function loadConfigEndpoints() {
-  const cache = getLogStores(CacheKeys.CONFIG_STORE);
-  let customConfig = await cache.get(CacheKeys.CUSTOM_CONFIG);
-
-  if (!customConfig) {
-    customConfig = await loadCustomConfig();
-  }
+  const customConfig = await getCustomConfig();
 
   if (!customConfig) {
     return {};
